player rotation implemented, as well as the bullet related rotation

+++ TrabalhoCIgrauA/Vertex.h
@@ -2,10 +2,10 @@
 class Vertex
 {
 public:
-	float pos_x;
-	float pos_y;
+	double pos_x;
+	double pos_y;
 
-	Vertex(float pos_x, float pos_y) {
+	Vertex(double pos_x, double pos_y) {
 		this->pos_x = pos_x;
 		this->pos_y = pos_y;
 	}
